feat: Initial logging version that builds
TODO:
* Test on Android 11 device (or emulator?)
