+++ GPSTest/build.gradle
@@ -126,6 +126,10 @@ dependencies {
     // QR Code reader for ground truth locations
     implementation 'com.google.zxing:android-integration:3.3.0'
 
+    // Write GNSS antenna info to JSON
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.11.2'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.11.2'
+
     // Map (Google flavor only)
     googleImplementation 'com.google.android.gms:play-services-maps:16.1.0'
     // Calculating offset for camera target in map view (Google flavor only)
