+++ GPSTest/src/main/java/com/android/gpstest/util/UIUtils.java
@@ -47,7 +47,8 @@ import androidx.fragment.app.Fragment;
 import com.android.gpstest.Application;
 import com.android.gpstest.BuildConfig;
 import com.android.gpstest.R;
-import com.android.gpstest.io.FileLogger;
+import com.android.gpstest.io.CsvFileLogger;
+import com.android.gpstest.io.JsonFileLogger;
 import com.android.gpstest.model.GnssType;
 import com.google.android.material.button.MaterialButton;
 import com.google.android.material.chip.Chip;
@@ -492,13 +493,13 @@ public class UIUtils {
      * @param activity
      * @param location
      * @param loggingEnabled true if logging is enabled, false if it is not
-     * @param fileLogger the file logger being used to log files
+     * @param csvFileLogger the file logger being used to log files
      * @param alternateFileUri The URI for a file if a file other than the one current used by the FileLogger should be used (e.g., one previously picked from the folder browse button), or null if no alternate file is chosen and the file from the file logger should be shared.
      * @return a dialog for sharing location and files
      */
     public static Dialog createShareDialog(AppCompatActivity activity, final Location location,
-                                           boolean loggingEnabled, FileLogger fileLogger,
-                                           Uri alternateFileUri) {
+                                           boolean loggingEnabled, CsvFileLogger csvFileLogger,
+                                           JsonFileLogger jsonFileLogger, Uri alternateFileUri) {
         View view = activity.getLayoutInflater().inflate(R.layout.share, null);
         AlertDialog.Builder builder = new AlertDialog.Builder(activity)
                 .setTitle(R.string.share)
@@ -631,7 +632,7 @@ public class UIUtils {
             }
         });
 
-        final File file = fileLogger.getFile();
+        final File file = csvFileLogger.getFile();
 
         if (loggingEnabled && file != null) {
             // Hide the logging instructions - logging is enabled and working
@@ -664,7 +665,7 @@ public class UIUtils {
 
         logBrowse.setOnClickListener(v -> {
             // File browse
-            Uri uri = IOUtils.getUriFromFile(activity, fileLogger.getFile());
+            Uri uri = IOUtils.getUriFromFile(activity, csvFileLogger.getFile());
             Intent intent = new Intent(Intent.ACTION_GET_CONTENT);
             intent.setData(uri);
             activity.startActivityForResult(intent, PICKFILE_REQUEST_CODE);
@@ -676,7 +677,7 @@ public class UIUtils {
             // Send the log file
             if (alternateFileUri == null) {
                 // Send the log file currently being logged to by the FileLogger
-                fileLogger.send(activity);
+                IOUtils.sendLogFile(activity, csvFileLogger, jsonFileLogger);
             } else {
                 // Send the log file selected by the user using the File Browse button
                 IOUtils.sendLogFile(activity, alternateFileUri);
