+++ GPSTest/src/main/res/values/strings.xml
@@ -291,12 +291,13 @@
     <string name="pref_file_location_output_title">Location</string>
     <string name="pref_file_location_output_summary">Logs location fix details to a file</string>
     <string name="pref_file_antenna_output_title">Antenna info</string>
-    <string name="pref_file_antenna_output_summary">Logs GNSS antenna details to a file</string>
+    <string name="pref_file_antenna_output_summary">Logs GNSS antenna details to a file. Only available on supported devices with Android 11 or greater.</string>
 
     <string name="logging_to_new_file">Logging to new file: %1$s</string>
     <string name="unable_to_close_all_file_streams">Unable to close all file streams.</string>
     <string name="could_not_initialize_file">Count not initialize file: %1$s</string>
     <string name="error_writing_file">Problem writing to file.</string>
+    <string name="unable_to_open_json_generator">Unable to open JSON generator</string>
 
     <string name="pref_about_category_title">About</string>
     <string name="pref_show_tutorial_title">Show tutorial</string>
